refactor section_header into section module
it's pretty clear that there is a good chunk of functionality
to implement for sections. I think it makes sense to refactor
this into a more general module so I can add the extra stuff
for different types of sections in a cleaner way.

Not really sure about my choice of putting the enums in its
own submodule in sections, may end up walking that back if
it seems like the wrong direction.